fd封装tcpconnection 设置watcher 纳入管理，watcher管理事件 切换编译器为clang

+++ xrtcserver/rtcserver/src/signaling/signaling_server.cpp
@@ -238,7 +238,7 @@ namespace xrtc {
     void SignalingServer::dispatch_conn(int fd)
     {
         int index=next_worker_++;
-        if(next_worker_>=workers_.size())
+        if(size_t(next_worker_)>=workers_.size())
         {
             next_worker_=0;
         }
